==> r/1-web-scraping.Rmd <==
---
title: "Muffin vs Cupcake: Web Scraping"
output: html_notebook
author: "Eric R. Scott"
---
This notebook will be for scraping the recipes ONLY.  At the end it should export a list of dataframes as a .rds file. (exporting the raw html doesn't work).
```{r packages, message=FALSE, warning=FALSE}
library(rvest)
#library(purrr)
#library(stringr)
library(tidyverse)
library(beepr) #for notification sounds when long functions are done running
#library(stringi)
```

# To Do:

```{r helpers}
#create function to read_html slowly to avoid getting banned
read_html_slow <- function(x, ...){
  output <- read_html(x)
  Sys.sleep(5)
  return(output)
}

read_html_safely <- possibly(read_html_slow, NA)
#testing:
#read_html_slow("http://allrecipes.com/recipe/233538/really-real-strawberry-cupcakes/")
```


# Get links to all cupcake and muffin recipes
allrecipes.com has a tag structure that is easy to navigate.  All the recipes for a category show up, but on several pages.  I didn't bother figuring out how many pages there were programatically, I just looked on the website.
```{r}
#This link takes you to page 1 by default
base_url <- "http://allrecipes.com/recipes/377/desserts/cakes/cupcakes/"
pages <- 1:12
cupcakes_urls <- paste0(base_url, "?page=", pages)

# map the read_html function to read all the pages at once
cupcakes <- map(cupcakes_urls, read_html_slow)
beep(4)

# extract the recipe links from all the pages
results <- cupcakes %>% map(~html_nodes(.,".fixed-recipe-card__title-link"))
cupcakes_links <- results %>% map(~html_attr(.,"href")) %>% flatten_chr()

# remove NAs
cupcakes_links <- cupcakes_links[!is.na(cupcakes_links)]
head(cupcakes_links)
```
Sweet! (literally) Now we have `r length(cupcakes_links)` cupcake recipes.

# Get links to muffin recipes
```{r}
base_url2 <- "https://www.allrecipes.com/recipes/350/bread/quick-bread/muffins/"
pages2 <- 1:5
muffin_urls <- paste0(base_url2, "?page=", pages2)


muffins <- map(muffin_urls, read_html_slow)
beep(4)

# extract the recipe links from all the pages
results2 <- muffins %>% map(~html_nodes(.,".fixed-recipe-card__title-link"))
muffin_links <- results2 %>% map(~html_attr(.,"href")) %>% flatten_chr()

# remove NAs
muffin_links <- muffin_links[!is.na(muffin_links)]
head(muffin_links)
```

# Scrape recipes!
To figure out if cupcakes are different than muffins, I need to get all the ingredients, categorize them as variables, get all the units the same for each variable across recipes, and exclude frosting or toppings. (I only care if the cake part is different from a muffin, obviously cupcakes are frosted and muffins are not).  I should standardize ingredient amounts by serving size.  Serving size and calorie count are in the same spot in each recipe, so I might as well get those too.

**Note:** For debugging, use the first code chunk below (default) that only uses a random sample of recipes.  To get all cupcake and muffin recipes, use the second code chunk
```{r sample}
#get random sample of recipe links
n = 26
cupcake_sample <- sample(cupcakes_links, n/2)
muffin_sample <- sample(muffin_links, n/2)
```

```{r all-recipes, eval=FALSE, include=FALSE}
#get ALL recipes!!!
cupcake_sample <- cupcakes_links
muffin_sample <- muffin_links
```


## Read in all the html of all the recipes
```{r read-recipes}
pb <- progress_estimated(length(cupcake_sample)) #makes a progress bar
cupcakes_html <- cupcake_sample %>%
  map(~{
    pb$tick()$print() #updates progress bar
    read_html_safely(.x)
    })

# anyNA(cupcakes_html)
# length(cupcakes_html[is.na(cupcakes_html)])

pb <- progress_estimated(length(muffin_sample))
muffin_html <- muffin_sample %>%
  map(~{
    pb$tick()$print() #updates progress bar
    read_html_safely(.x)
    })
beep(4)
# anyNA(muffin_html)
# length(muffin_html[is.na(muffin_html)])
```

## Filter out any that didn't work
If a recipe didn't get read in, I want to remove that link from `cupcake_sample` and `muffin_sample` because I'll use the same links later on to extract calories and servings, then remove it from `cupcakes_html` and `muffin_html` so that `html_nodes()` works on it.

```{r}
#first the links
# cupcake_sample[is.na(cupcakes_html)]
cupcake_sample <- cupcake_sample[!is.na(cupcakes_html)]
muffin_sample <- muffin_sample[!is.na(muffin_html)]

#then the html
cupcakes_html <- cupcakes_html[!is.na(cupcakes_html)]
muffin_html <- muffin_html[!is.na(muffin_html)]
```


## Extract ingredients and categorize into "parts"
Most cupcake recipes divide the ingredients list with headers that end in a colon, like "Cake:" or "Frosting:".  I can use that to create a variable, `part`, that later will help me filter out toppings and frostings.  I doubt muffins will have "parts" that are useful.  They might be separated into wet and dry ingredients, but those are both part of the muffin.

```{r extract-parts}
cupcakes_ing <- cupcakes_html %>% map(~html_nodes(., ".added")) %>% map(html_text)

cupcakes_ing <- cupcakes_ing %>% map(~ as_tibble(.) %>%
                       mutate(part = str_extract(.$value, ".+:$")) %>% 
                       fill(part) %>% 
                       filter(!str_detect(.$value, ":")) %>% 
                       filter(value != "Add all ingredients to list"))
# head(cupcakes_ing)
```

## Identify cake vs. frosting
Only keep "parts" that have the word "cake" in them for cupcakes.
*What if there are no "parts"?* I think it is safe to assume that if a cupcake recipe has a "part" for some ingredients but not for others, the ingredients without a "part" are the cake.  If a cupcake recipe has no "parts" listed at all, then it is not safe to assume all ingredients are for the cake (see "problems" below for examples and more discussion).


```{r fill-in-parts}
cupcakes_ing <- cupcakes_ing %>% map(~
  mutate(., part = ifelse(any(!is.na(part)) & is.na(part), "Cake:", part)))
# take cupcakes_ing and map the following mutate function to it:
# For a given recipe, if any of the "parts" are not blank and the current ingredient's part *is* blank, then that ingredient's part is "Cake:", otherwise just keep the part you previously assigned it.
```
Muffins won't have the problem of "parts", but I should still create that column so I can `bind_rows()` later without an issue.

```{r}
muffins_ing <- muffin_html %>% map(~html_nodes(., ".added")) %>% map(html_text)

muffins_ing <- muffins_ing %>% map(~ as_tibble(.) %>%
                       mutate(part = str_extract(.$value, ".+:$")) %>% 
                       fill(part) %>% 
                       filter(!str_detect(.$value, ":")) %>% 
                       filter(value != "Add all ingredients to list"))
```

## Extract calories and serving size
Occasionally a recipe seems to not have calories.  I need to catch these and convert to NA
```{r}
cupcake_calories <- cupcakes_html %>% 
  map(~html_nodes(., ".calorie-count span:nth-child(1)")) %>%
  map(html_text) %>%
  map(., ~ifelse(identical(., character(0)), NA, .)) #replaces character(0) with NA

#cupcake_calories[[166]]

cupcake_servings <- cupcakes_html %>% 
  map(~html_nodes(., ".subtext")) %>%
  map(html_text) %>% 
  map(~str_extract(.,"(?<=yields\\s)\\d+")) %>% 
  map(., ~ifelse(identical(., character(0)), NA, .)) #replaces character(0) with NA

cupcakes_ing <- map2(cupcakes_ing, cupcake_servings, ~mutate(.x, servings = as.integer(.y)))
cupcakes_ing <- map2(cupcakes_ing, cupcake_calories, ~mutate(.x, calories = as.integer(.y)))

muffin_calories <- muffin_html %>% 
  map(~html_nodes(., ".calorie-count span:nth-child(1)")) %>%
  map(html_text) %>% 
  map(., ~ifelse(identical(., character(0)), NA, .)) #replaces character(0) with NA
# length(muffin_calories)

muffin_servings <- muffin_html %>% 
  map(~html_nodes(., ".subtext")) %>%
  map(html_text) %>% 
  map(~str_extract(.,"(?<=yields\\s)\\d+")) %>% 
  map(., ~ifelse(identical(., character(0)), NA, .)) #replaces character(0) with NA
# length(muffin_servings)

muffins_ing <- map2(muffins_ing, muffin_servings, ~mutate(.x, servings = as.integer(.y)))
muffins_ing <- map2(muffins_ing, muffin_calories, ~mutate(.x, calories = as.integer(.y)))

# length(muffins_ing)
```
# Name data frames
Add recipe ID as a name to each data frame.  This recipe ID can be used to generate a URL to the recipe.
```{r}
#add names to list of data frames
names(cupcakes_ing) <- str_extract(cupcake_sample, "(?<=/)\\d+")
names(muffins_ing) <- str_extract(muffin_sample, "(?<=/)\\d+")
```

# Export lists of tibbles
At this point, all the webscraping is done.  I'll export this to a .rds file and continue the data wrangling in another notebook.

```{r}
write_rds(muffins_ing, "muffin_raw.rds")
write_rds(cupcakes_ing, "cupcake_raw.rds")
```
