==> r/2-wrangling.Rmd <==
---
title: "Muffin vs Cupcake: Data Wrangling"
output: html_notebook
author: "Eric R. Scott"
---
```{r}
library(tidyverse)
# library(rvest)
library(stringi)
library(stringr)
library(here)
```
Take the html from the cupcake and muffin recipes and convert it into a data frame/tibble.

# To-Do:

- Figure out why so many recipes are missing by the end. Cupcakes go from ~200 to 45 after removing ones with ambiguous "part" assignments for ingredients.  The "Flatten" step loses 3 muffins for some unkown reason.
- Filter out recipes with no flour.  Seems like that's probably not actually a muffin or cupcake. Must be done after `spread()`ing since recipes with no flour won't list as ingredient (can't use `filter()` easily)

# Problems:

1. **Recipes without defined "parts"** (DONE)

Some recipes don't define what ingredients are for the cakes vs. the frosting.  My intention is to exclude frosting ingredients from this analysis, so the safest way to deal with this is to assume that unless it has the word "cake" somewhere in the `part` column, then the ingredient is not for the cake part.

- Some have a "part" for frosting, but not for the cake. (e.g. 261244, 255301)  So in that case, all the ingredients that are NA for `part` go with the cake.

- Some don't have frosting or toppings, and all the ingredients are for the cake (e.g. 235291, 239265).

- Others list all the ingredients for the frosting and cake in one list (e.g. 245175, 218356).

- Yet others are complete bullshit and have "cupcakes" as one of the ingredients (261787)

So far I've dealt with this by assuming if a recipe has any parts, and some ingredients have no associated part, then they are for the cake.  (see code chunk 10: fill-in-parts).

2. **Weird ingredients with non-standard units**

Some recipes have ingredients without standard units like "2 apples".  I can imagine several ways of dealing with these ingredients, with different consequences

 - Remove Entire Recipes: This would leave only recipes with measured ingredients.  It would err on the side of removing recipes with things like "cooking spray" to grease the tins that don't really affect batter volume.  However, the benefit is that I could add up ingredients to get a total batter volume.  I'd then have my choice of converting ingredient amounts to per serving *or* per total volume (percentage).
 
 - Remove Problem Ingredients: I could also remove/ignore ingredients without units, but keep the rest of the recipe.
 
 - **Tally the weird shit up:** I could create a category for "unitless" ingredients and just count how many there are in each recipe.  This is my current strategy.

3. **Cheaters**

Several recipes use Bisquick, cake mix, or even *store-bought cupcakes* as ingredients.  These recipes are obviously cheating and should be removed!!  One way to do this might be to require recipes to all have flour.  I should check to see if any have "self-rising flour".

# Helper functions

```{r}
str_detect_any <- function(string, pattern){
  map_lgl(string, ~stri_detect_regex(., pattern) %>% any(.))
}
```


# Read in lists of data frames

```{r}
muffins_ing <- read_rds(here("data", "muffin_raw.rds"))
cupcakes_ing <- read_rds(here("data", "cupcake_raw.rds"))
```

# Flatten and Bind

Use `bind_rows()` to create one data frame for all recipes in `cupcakes_ing` and one for all recipes in `muffins_ing` using recipe ID from the data frame names as the id. Then `bind_rows()` of muffins and cupcakes using "muffin" or "cupcake" as the id.

```{r}
cupcakes_df <- cupcakes_ing %>% bind_rows(.id = "recipe_id")
muffins_df <- muffins_ing %>% bind_rows(.id = "recipe_id")
# muffins_df %>% group_by(recipe_id) %>% summarise(n()) %>% nrow()
```

```{r bind-all-recipes}
#add "cupcake" or "muffin" as name
recipes <- bind_rows(cupcakes_df, muffins_df, .id = "type") %>% mutate(type = ifelse(type == 1, "cupcake", "muffin"))
recipes
#write_rds(recipes, "recipes_raw.rds")
```

# Extract Ingredient Variables

I'll use regular expressions to try to extract the amount and units.  Then in the same `mutate()` call I'll try to categorize ingredients into discrete variables.

First I define some ingredients that will be grouped.  Further grouping of ingredients may be done later.

```{r ingredient-groups}
fruits <- c("\\w*apple", "\\w*berr(y|ies)", "raisin", "lemon", "banana",
            "orange", "pear", "peach", "rhubarb", "plum", "cherry", "fruit", "lime", "mango", "currant", "cherr(y|ies)")

nuts <- c("\\w*nut", "almond", "pecan", "seed")

veggies <- c("carrot", "pumpkin", "potato", "zucchini", "\\bcorn\\b", "\\byam")

spices <- c("cinnamon", "nutmeg", "clove", "ginger", "spice", "cardamom", "pepper")

low_cal <- c("stevia", "splenda", "sucralose", "truvia", "erythritol")

trash <- c("paper", "reynolds", "cooking spray", "cutter", "baking spray")
```


```{r}
recipes2 <- recipes %>%  
  mutate(value = tolower(value)) %>% 
  mutate(amount = str_extract(.$value, "^[\\d, /]+"),
         units = str_extract(.$value, "(\\(.+\\)|[^0-9, /]+)"), #finds either something in parens, or next word after numbers
         ingredient = case_when(#cheaters
                                str_detect(.$value, "\\bmix\\b")    ~ "mix",
                                str_detect(.$value, "cupcake")      ~ "cupcakes",
                                #sweeteners
                                str_detect(.$value, "sugar")        ~ "sugar",
                                str_detect(.$value, "juice")        ~ "fruit juice",
                                str_detect(.$value, "honey")        ~ "honey",
                                str_detect(.$value, "agave")        ~ "agave",
                                str_detect_any(.$value, low_cal)    ~ "low-cal sweetener",
                                str_detect(.$value, "syrup")        ~ "syrup",
                                str_detect(.$value, "molasses")     ~ "molasses",
                                str_detect(.$value, "frosting")     ~ "frosting",
                                #starches
                                str_detect(.$value, "flour")        ~ "flour",
                                str_detect(.$value, "starch")       ~ "starch",
                                str_detect(.$value, "corn\\s?meal") ~ "cornmeal",
                                str_detect(.$value, "bran")         ~ "bran",
                                str_detect(.$value, "\\boat")       ~ "oats", # oat but not "coated"
                                str_detect(.$value, "wheat germ")   ~ "wheat germ",
                                #dairy
                                str_detect(.$value, "buttermilk")   ~ "buttermilk", #snag before butter and milk
                                str_detect(.$value, "milk")         ~ "milk",
                                str_detect(.$value, "cream cheese") ~ "cream cheese",
                                str_detect(.$value, "sour cream")   ~ "sour cream",
                                str_detect(.$value, "tartar")       ~ "other", #snag cream of tartar before cream
                                str_detect(.$value, "\\bcream\\b")  ~ "cream", #not "creamy"
                                str_detect(.$value, "yogurt")       ~ "yogurt",
                                str_detect(.$value, "cheese")       ~ "cheese", #cheeses that are not cream cheese
                                #fats
                                str_detect(.$value, "\\boil\\b")    ~ "oil", #not "boiling"
                                str_detect(.$value, "shortening")   ~ "shortening",
                                str_detect(.$value, "mayonnaise")   ~ "mayonnaise",
                                str_detect(.$value, "margarine")    ~ "margarine",
                                  #butter has to go at the end, after apple butter and peanut butter are captured
                                #leavening
                                str_detect(.$value, "baking soda")  ~ "baking soda",
                                str_detect(.$value, "baking powder")~ "baking powder",
                                #other
                                str_detect(.$value, "\\bsalt\\b")   ~ "salt", #not "salted" or "unsalted"
                                str_detect(.$value, "water")        ~ "water",
                                str_detect(.$value, "vanilla extract")  ~ "vanilla",
                                str_detect(.$value, "chocolate")    ~ "chocolate",
                                str_detect(.$value, "cocoa")        ~ "chocolate",
                                str_detect(.$value, "egg")          ~ "eggs",
                                str_detect(.$value, "vinegar")      ~ "vinegar",
                                #spices
                                str_detect_any(.$value, spices)     ~ "spice",
                                #nuts
                                str_detect_any(.$value, nuts)       ~ "nut",
                                #fruit
                                str_detect_any(.$value, fruits)     ~ "fruit",
                                #veggies
                                str_detect_any(.$value, veggies)    ~ "vegetable",
                                #butter
                                str_detect(.$value, "butter")       ~ "butter",
                                str_detect_any(.$value, trash)      ~ "trash",
                                TRUE ~ "other"))

recipes2 %>% select(value, ingredient, units, amount, everything()) %>% filter(ingredient == "trash")
#filter out trash
recipes2 <- recipes2 %>% filter(ingredient != "trash")
```

After testing this with several different random samples of cupcake recipes, I'm pretty happy with it. 

Things I've noticed that are wrong/missed:

- "creamy natural peanut butter" is categorized as cream
- coffee
- whiskey(?!?)
- "caramel ice cream topping" is categorized as cream
- tea
- bacon grease
- xanthan gum
- bisquick

# Deal with cheaters
Get rid of recipes that call for self-rising flour, bisquick, cupcakes, or cake mix.  I *could* figure out what self-rising flour is, but that seems like too much work.

```{r}
cheaters <- recipes2 %>%
  filter(str_detect_any(value, c("self-rising", "bisquick", "cupcake", "mix\\b"))) %>%
  .$recipe_id

recipes2 <- recipes2 %>% filter(!recipe_id %in% cheaters)
#remove recipeID's that match this
length(cheaters)
```

# Convert fractions to decimals
The `ammount` column is character class.  I want numeric.

I think one approach would be to turn a space into a "+" then use mutate to do `eval(parse(ammount))`

```{r}
recipes3 <- recipes2 %>% 
  mutate(amount = str_trim(amount)) %>% 
  mutate(dec_amount = str_replace(amount, "\\s", "+")) %>% 
  mutate(dec_amount = sapply(dec_amount, function(x) eval(parse(text=x))))
recipes3 %>% select(amount, dec_amount, everything())
```

## What gets missed?

```{r}
recipes3 %>% filter(is.na(dec_amount))
```

"pinch" gets missed.
Fixed:

```{r}
recipes3 <- recipes3 %>%
  mutate(dec_amount = ifelse(units == "pinch" & is.na(amount), 1, dec_amount))
```

# Unit conversions

## Deal with package size

Occasionally an ingredient is like "1 (8 ounce) packet of cream cheese" and my script so far detects "1" as the amount and "(8 ounce)" as the units.  I should deal with this.

**Deal with "oz." and possible decimal points (e.g. "12.5 ounce")** (e.g. 234135)

```{r}
recipes3 <- recipes3 %>%
  mutate(dec_amount = ifelse(str_detect(units, "\\(.+ (ounce|oz.)"),
                          as.numeric(str_extract(units, "\\.?\\d+\\.?\\d*")) * dec_amount,
                          dec_amount),
         units = ifelse(str_detect(units, "\\(.+ (ounce|oz.)"), "ounces", units))
```

```{r}
#overwrite original amount column to avoid confusion
recipes4 <- recipes3 %>%
  select(-amount) %>%
  rename(amount = dec_amount) %>% 
  select(type, ingredient, amount, units, servings, calories, value,recipe_id, part)
```

## Unit conversions
Just for the sake of uniformity, let's put everything in "cups" except eggs.

```{r}
#units we should be able to convert to "cups"
convertable <- c("tablespoon", "teaspoon", "cup", "ounce", "pint", "pinch")

recipes5 <- recipes4 %>% 
  mutate(amount_std = case_when(ingredient == "eggs"            ~ amount,
                                str_detect(units, "cup")        ~ amount,
                                str_detect(units, "pinch")      ~ amount / 768,
                                str_detect(units, "teaspoon")   ~ amount * 0.0208333,
                                str_detect(units, "tablespoon") ~ amount * 0.0625,
                                str_detect(units, "ounce")      ~ amount * 0.125,
                                str_detect(units, "pint")       ~ amount * 2,
                                TRUE                            ~ as.numeric(NA))) %>%
  #change units to cups
  mutate(units_std = case_when(str_detect_any(units, convertable) ~ "cups",
                               ingredient == "eggs"               ~ "eggs",
                               TRUE                               ~ as.character(NA))) %>% 

  select(recipe_id, ingredient, amount, units, amount_std, units_std, everything())
```

# Filter out recipes with no flour
There's gotta be something wrong if a recipe has no flour in it.

```{r}
flourless <- recipes5 %>%
  group_by(recipe_id) %>%
  summarise(hasflour = any(ingredient == "flour")) %>%
  filter(hasflour == FALSE) %>% 
  .$recipe_id

recipes5a <- recipes5 %>%
  filter(!recipe_id %in% flourless)
```

They're all mini cheesecakes I bet.  That is **not** a cupcake

# Collapse ingredients
Some recipes may have multiple ingredients within an ingredient category.  For example, a recipe might have two kinds of fruit, or it might have 2 eggs plus 2 eggwhites, or two different kinds of sugar (granulated and brown).  I need to add up those amounts so each variable is only listed once per recipe.

At the same time, I might be able to deal with ingredients with no units in some way.  These are mostly going to be ingredients like "2 apples."  It's not practical to estimate the volume of every one of these ingredients, but it seems "wrong" to ignore them.

I see three options

1. Ignore these ingredients.  Just remove ingredients with an `NA` for `units_std`
2. Remove problem recipes.  Remove recipes with any ingredient that has `NA` for `units_std`
3. **Count these ingredients.**  Instead of adding up actual amounts when summarising, convert the `amount_std` to 1 for all ingredients with `unit_std == NA`.  Then the sum will represent a count rather than a volume.  These ingredients should also be flagged as "unitless" so they can be removed from particular analyses if count data are inappropriate.


```{r}
#set up to "count" unitless ingredients
recipes5b <- recipes5a %>% 
  mutate(ingredient = ifelse(is.na(units_std), "unitless", ingredient),
         amount_std = ifelse(is.na(units_std), 1, amount_std))

recipes_tidy <- recipes5b %>% 
  group_by(type, recipe_id, ingredient) %>% 
  summarise(amount_sum = sum(amount_std),
            units_std = first(units_std),
            servings = first(servings),
            calories = first(calories),
            value = paste(value, collapse = ";")
  )

recipes_tidy
recipes_tidy %>% filter(ingredient == "unitless")
recipes_tidy %>% filter(ingredient == "other")
```
There are actually a lot of "unitless" ingredients, but most are in frosting I think?

# Calculate amount per serving
Normalize by serving amount

```{r}
recipes_tidy <-
  recipes_tidy %>%
  mutate(amount_per = amount_sum/servings,
         calories_per = calories/servings)
```


# Write full, tidied dataset to RDS
This data set is basically ready for analysis, although it contains cupcake recipes with frosting (part != cake) or with unclear grouping of ingredients (all parts = NA).

```{r}
write_rds(recipes_tidy, here("data", "recipes_tidy.rds"))
write_csv(recipes_tidy, here("data", "recipes_tidy.csv"))
```


# Filter out frosting

We all know cupcakes have frosting and muffins don't, so let's create a subset of `recipes_tidy` that removes frosting.  This code keeps all the muffin recipes intact, but only keeps cupcake ingredients that are definitely part of the cake part of a cupcake.  This results in some cupcake recipes being elminated when the "part" the ingredients are for is ambiguous.

```{r}
nofrosting_tidy <-
  recipes5b %>%
  #all ingredients that are from muffin recipes or cupcake recipes where the part is definitely cake and not frosting
  filter(type == "muffin" | type == "cupcake" & str_detect_any(tolower(part), c("cake", "dry", "wet", "batter")) & !str_detect_any(tolower(part), c("decoration", "garnish", "frosting", "icing"))) %>% 
  #This code is the same code that created recipes_tidy above
  group_by(type, recipe_id, ingredient) %>% 
  summarise(amount_sum = sum(amount_std),
            units_std = first(units_std),
            servings = first(servings),
            calories = first(calories),
            value = paste(value, collapse = ";")
  ) %>% 
    mutate(amount_per = amount_sum/servings,
         calories_per = calories/servings)
nofrosting_tidy
```

This subset has about half the number of rows as the full dataset.  That's a lot of frosting/ambiguous ingredients!

I should remove the "calories" column since they no longer reflect the ingredients left.

```{r}
nofrosting_tidy <- 
  nofrosting_tidy %>%
  select(-calories, - calories_per)
```

## Write frostingless dataset to RDS

```{r}
write_rds(nofrosting_tidy, here("data", "nofrosting_tidy.rds"))
write_csv(nofrosting_tidy, here("data", "nofrosting_tidy.csv"))
```
